examples_requirements.txt
django-storages==1.1.8
django-boto==0.3.9
